+++ include/Gaffer/TypeIds.h
@@ -146,6 +146,8 @@ enum TypeId
 	ContextQueryTypeId = 110099,
 	TweakPlugTypeId = 110100,
 	TweaksPlugTypeId = 110101,
+	V2fVectorDataPlugTypeId = 110102,
+	V3iVectorDataPlugTypeId = 110103,
 
 	LastTypeId = 110159,